:arrow_up: Update pyfairylandfuture dependency, modify import paths and improve code
This commit updates pyfairylandfuture to version 1.1.0 and modifies the import paths accordingly. It also includes improvements to settings, journal, configuration and model fields for enhanced code clarity and accuracy. Furthermore, verbose details have been added to the serializer, along with the addition of trusted hosts for better package installation settings.